+++ lib/rails-i18nterface.rb
@@ -6,7 +6,6 @@ require 'rails-i18nterface/cache'
 require 'rails-i18nterface/yamlfile'
 require 'rails-i18nterface/sourcefiles'
 require 'rails-i18nterface/keys'
-require 'rails-i18nterface/storage'
 
 module RailsI18nterface
 end