+++ spec/internal/app/controllers/application_controller.rb
@@ -1,6 +1,8 @@
 # encoding: utf-8
 
 class ApplicationController < ActionController::Base
+
   def index
   end
+
 end