+++ spec/lib/keys_spec.rb
@@ -11,36 +11,36 @@ describe RailsI18nterface::Keys do
     @keys = RailsI18nterface::Keys.new(@root_dir, :sv, :no)
   end
 
-  describe 'to_a' do
-    it 'extracts keys from I18n lookups in .rb, .html.erb, and .rhtml files' do
-      @keys.files.keys.map(&:to_s).sort.should == [
-        'activerecord.attributes.article.active',
-        'activerecord.attributes.article.body',
-        'activerecord.attributes.article.created_at',
-        'activerecord.attributes.article.title',
-        'activerecord.attributes.article.updated_at',
-        'activerecord.attributes.topic.created_at',
-        'activerecord.attributes.topic.title',
-        'activerecord.attributes.topic.updated_at',
-        'activerecord.models.article',
-        'activerecord.models.topic',
-        'article.key1',
-        'article.key2',
-        'article.key3',
-        'article.key4',
-        'article.key5',
-        'category_erb.key1',
-        'category_html_erb.key1',
-        'category_rhtml.key1',
-        'js.alert'
-      ]
-    end
+  it 'extracts keys from I18n lookups in .rb, .html.erb, and .rhtml files' do
+    @keys.files.keys.map(&:to_s).sort.should == [
+      'activerecord.attributes.article.active',
+      'activerecord.attributes.article.body',
+      'activerecord.attributes.article.created_at',
+      'activerecord.attributes.article.title',
+      'activerecord.attributes.article.updated_at',
+      'activerecord.attributes.topic.created_at',
+      'activerecord.attributes.topic.title',
+      'activerecord.attributes.topic.updated_at',
+      'activerecord.models.article',
+      'activerecord.models.topic',
+      'article.key1',
+      'article.key2',
+      'article.key3',
+      'article.key4',
+      'article.key5',
+      'category_erb.key1',
+      'category_html_erb.key1',
+      'category_rhtml.key1',
+      'js.alert'
+    ]
   end
 
-  describe 'to_hash' do
-    it 'return a hash with I18n keys and file lists' do
-      @keys.files[:'article.key3'].should == ['app/models/article.rb']
-    end
+  it 'return a hash with I18n keys and file lists' do
+    @keys.files[:'article.key3'].should == ['app/models/article.rb']
+  end
+
+  it 'reloads the translatable strings' do
+    @keys.reload(@root_dir).size.should == 123
   end
 
   describe 'i18n_keys' do